Generated by Spark: I would like to add notification settings within the settings app, an appearance setting with seperate toggles for light mode, dark mode or system default add edit profile functionality to be able to change avatar, upload profile picture, change display name. this should in turn also get rid of the camera icon of the profile avatar, since that option will now be in the edit profile button. Home should be the center button with schedule and profile on the right and and training and library on the left of that button, I would like the home button to be circular and I would like this bottom navigation bar to be red with the different navigation buttons to be white background but red lettering
when signing on, the district code should only be asked if the system recognizes the user as it being their first time logging in. other than that, it should then that it should take them directly to their dashboard once it goes through the sso process. Wether it's log in with google, or log in with apple, etc.

I also want there to be a thin red border that encompasses a very thin margin around the screen. i don't want the users to see a scroll bar and make sure the name tag updates with any screen name changes they might make in their profile