Generated by Spark: The developer console should be the home screen for anyone who is a developer. there is no need to have the typical home screen like for other users. the owners hoemscreen would be the owners console as well. see, pretty much everyone aside from trainees and teammbers wont actually be doing trainings. so only if a trainnig is active for them should it show up on their dashboard.
on the login popup showing wether or not the user is on web or mobile is unneccesary for login, but it would be a good metric to see for the opwners and devs for how people are accessing the app.

the "sign in to access your training dashboard"  message in the sign in popup can disappear. it's unneeded since we have a sign in button. also, there should be 3 chopices to sign in. one should be just sign in, and then one should be sign in with google and one should be sign in wiht apple.

here is some info from google about how to integrate: The easiest way to add a Google Sign-In button to your site is to use an automatically rendered sign-in button. With only a few lines of code, you can add a button that automatically configures itself to have the appropriate text, logo, and colors for the sign-in state of the user and the scopes you request.

To create a Google Sign-In button that uses the default settings, add a div element with the class g-signin2 to your sign-in page:

<div class="g-signin2" data-onsuccess="onSignIn"></div>
Get profile information
After you have signed in a user with Google using the default scopes, you can access the user's Google ID, name, profile URL, and email address.

To retrieve profile information for a user, use the getBasicProfile() method.

function onSignIn(googleUser) {
  var profile = googleUser.getBasicProfile();
  console.log('ID: ' + profile.getId()); // Do not send to your backend! Use an ID token instead.
  console.log('Name: ' + profile.getName());
  console.log('Image URL: ' + profile.getImageUrl());
  console.log('Email: ' + profile.getEmail()); // This is null if the 'email' scope is not present.
}

Important: Do not use the Google IDs returned by getId() or the user's profile information to communicate the currently signed in user to your backend server. Instead, send ID tokens, which can be securely validated on the server.
Sign out a user
You can enable users to sign out of your app without signing out of Google by adding a sign-out button or link to your site. To create a sign-out link, attach a function that calls the GoogleAuth.signOut() method to the link's onclick event.

<a href="#" onclick="signOut();">Sign out</a>
<script>
  function signOut() {
    var auth2 = gapi.auth2.getAuthInstance();
    auth2.signOut().then(function () {
      console.log('User signed out.');
    });
  }
</script>